refactor: useExplosion 커스텀훅 분리
